Feature/151 option auto approach

+++ main/midnight/src/panels/panel_think.h
@@ -42,7 +42,7 @@ protected:
 
 protected:
     flags32                 flags;
-    s64                     currentPage;
+    s32                     currentPage;
     mxid                    id;
     mxid                    objectId;
     int                     current_tab;
